Change index type from int to int64
To avoid the int overflow of 0x80000000 of 386 arch

+++ src/wallet/bip44wallet/account_test.go
@@ -51,7 +51,7 @@ func TestNewBip44Account(t *testing.T) {
 	tt := []struct {
 		name           string
 		accountName    string
-		index          int
+		index          int64
 		seed           string
 		seedPassphrase string
 		coinType       wallet.CoinType
